Enable setting security contexts on rootfs inodes.
rootfs (ramfs) can support setting of security contexts
by userspace due to the vfs fallback behavior of calling
the security module to set the in-core inode state
for security.* attributes when the filesystem does not
provide an xattr handler.  No xattr handler required
as the inodes are pinned in memory and have no backing
store.

This is useful in allowing early userspace to label individual
files within a rootfs while still providing a policy-defined
default via genfs.

Change-Id: I866bd3a8ed626d82007a2108ede978dd6e5a189b
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>
Signed-off-by: CallMESuper <zj7859256@gmail.com>